+++ backend/api/routes/auth.ts
@@ -1,12 +1,10 @@
 import express, { Request, Response } from 'express';
-import { login, register } from '../controllers/authControllers';
+import { login, register, deleteAccount } from '../controllers/authControllers';
 import passport from 'passport';
 import cors from 'cors';
 import jwt, { JwtPayload } from 'jsonwebtoken';
 import env from '../config/env';
 import { authenticate } from '../middleware/auth';
-import RecipeList from '../models/recipeList';
-import User from '../models/user';
 
 console.log('Auth routes file loaded');
 
@@ -128,34 +126,7 @@ router.post('/logout', (req: Request, res: Response) => {
   });
 });
 
-router.delete('/delete-account', authenticate, async (req: Request, res: Response) => {
-  try {
-    const userId = (req.user as any).id;
-
-    const deletedUser = await User.findByIdAndDelete(userId);
-
-    if (!deletedUser) {
-      res.status(404).json({ error: 'User not found' });
-      return;
-    }
-
-    await RecipeList.deleteMany({ user: userId });
-
-    res.status(200).json({ message: 'Account deleted successfully' });
-  } catch (error) {
-    console.error('Error deleting account:', error);
-    res.status(500).json({ error: 'Failed to delete account' });
-  }
-});
-
-router.route('/delete-account-alt').delete(authenticate, async (req, res) => {
-  try {
-    const userId = (req.user as any).id;
-    res.status(200).json({ message: 'Account deleted via alt route' });
-  } catch (error) {
-    res.status(500).json({ error: 'Failed to delete via alt route' });
-  }
-});
+router.delete('/delete-account', authenticate, deleteAccount);
 
 console.log(
   'Auth routes registered:',
